Add files via upload
The objective of this project is to predict the closing price of a company's stock at the end of the next trading day. By leveraging a combination of text and structured data analysis, the project aims to develop a robust machine-learning model capable of accurate predictions.

The dataset consists of 1,300 trading days, with the closing prices for the next trading day provided for the first 1,000 days. Using the complete data for these 1,000 days, a machine learning model will be trained to forecast the closing price for the remaining 300 trading days where the next day's closing price is not provided.